backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
anthropic==0.7.8
openrouter==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
